pyproject.toml
[project]
name = "nonebot-plugin-skland"
version = "0.1.0"
description = "通过森空岛查询游戏数据"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "nonebot-plugin-alconna>=0.54.2",
    "nonebot2>=2.4.1",
]

[project.urls]
homepage = "https://github.com/FrostN0v0/nonebot-plugin-skland"
repository = "https://github.com/FrostN0v0/nonebot-plugin-skland"

[project.optional-dependencies]
dev = [
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot2[fastapi]>=2.4.1",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.nonebot]
plugins = ["nonebot_plugin_skland"]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]


[tool.isort]
profile = "black"
line_length = 90
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
line-length = 90
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901"]

[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"
typeCheckingMode = "basic"
